Fix for #55: `CosineDist` fails with Ints
The reason CosineDist was failing for integers was because
`eval_start` was only defined for subtypes of AbstractFloat instead
of Number. Since `zero` has a method in Base for any Number, it seemed
reasonable to widen the net on the types of arrays that `eval_start`
instead of writing a new method on Integers.