Add dims argument to pairwise
This allows specifying whether observations are stored as columns (dims=2) or as rows (dims=1).
All functions are still written internally to perform computations over columns,
computations over rows are supported by calling transpose on the matrices.

Add a deprecation to require specifying explicitly dims=1 or dims=2: contrary to this package,
other implementations most frequently default to dims=1, and even in Julia the cov and cor functions
default to dims=1.